Proposal: Defer Sentry reporting to raven-for-redux
Do not merge! This is just a proof of concept. If we like it, I will update
tests and the README to reflect this change.

redux-raven-middleware is actually a superset of raven-for-redux, so I figured
I would propose a merger. This would allow users of redux-raven-middleware to
gain some of the advantages of raven-for-redux without having to change any of
their code. It would also avoid having two libraries duplicating the same
effort.

Advantages
---

* Add state/last-action to Raven as context for all errors, not just reducer errors (#21)
* Option to add `data` to redux action breadcrumbs (#24)
* Users can now configure the breadcrumb category
* Less code to maintain

Disadvantages
---

* Logger feature is removed